[uss_qualifier] prepare SCD auth check scenario to check different endpoint groups
This PR prepares the `AuthenticationValidation` scenario for checking the endpoints for constraint references and USS availability states.

For those, the notion of a 'wrong' scope is different than for subscriptions and SCD, as the wront scope for SCD might be the correct one for setting a USS availability state.

(PRs introducing the checks for the other endpoints will follow – this is just a preparatory cleanup)